+++ cypress/component/Icon.spec.tsx
@@ -2,21 +2,23 @@ import React from 'react'
 import { Container, Typography } from '@toptal/picasso'
 import * as icons from '@toptal/picasso/Icon'
 
+const component = 'Icon'
+
 describe('Icon', () => {
   it('renders all icons', () => {
     cy.mount(
-      <Container
-        style={{
-          display: 'grid',
-          gridTemplateColumns: 'repeat(auto-fill, minmax(213px, 1fr))',
-        }}
-      >
+      <Container flex style={{ flexWrap: 'wrap' }}>
         {Object.entries(icons).map(([iconName, IconComponent]) => {
           return (
             <Container
               bordered
               padded='small'
-              style={{ overflow: 'hidden', textAlign: 'center' }}
+              style={{
+                overflow: 'hidden',
+                textAlign: 'center',
+                width: 213,
+                height: 80,
+              }}
             >
               <Typography>{iconName}</Typography>
               <IconComponent color='red' />
@@ -25,6 +27,10 @@ describe('Icon', () => {
         })}
       </Container>
     )
-    cy.get('body').happoScreenshot()
+
+    cy.get('body').happoScreenshot({
+      component,
+      variant: 'all-icons',
+    })
   })
 })