+++ packages/picasso-provider/src/Picasso/NotificationsProvider/NotificationsProvider.tsx
@@ -1,9 +1,8 @@
 import { makeStyles, Theme } from '@material-ui/core/styles'
-import cx from 'classnames'
 import { SnackbarProvider } from 'notistack'
 import React from 'react'
 
-import { useDrawer, usePageTopBar } from '../RootContext'
+import { usePageTopBar } from '../RootContext'
 import styles from './styles'
 
 const useStyles = makeStyles<Theme>(styles, {
@@ -21,7 +20,6 @@ interface Props {
 const NotificationsProvider = ({ children, container }: Props) => {
   const { hasTopBar } = usePageTopBar()
   const classes = useStyles()
-  const { hasDrawer } = useDrawer()
 
   const containerAnchorOriginTop = hasTopBar
     ? classes.rootWithMargin
@@ -32,9 +30,8 @@ const NotificationsProvider = ({ children, container }: Props) => {
       maxSnack={MAX_NOTIFICATION_MESSAGES}
       domRoot={container}
       classes={{
-        containerAnchorOriginTopRight: cx(containerAnchorOriginTop, {
-          [classes.marginWithDrawer]: hasDrawer,
-        }),
+        containerRoot: classes.root,
+        containerAnchorOriginTopRight: containerAnchorOriginTop,
         containerAnchorOriginTopLeft: containerAnchorOriginTop,
         containerAnchorOriginTopCenter: containerAnchorOriginTop,
       }}