+++ packages/picasso/src/FileInput/story/index.jsx
@@ -23,10 +23,7 @@ page.createTabChapter('Props').addComponentDocs({
 page
   .createChapter()
   .addExample('FileInput/story/Default.example.tsx', 'Default')
-  .addExample('FileInput/story/CustomButtonLabel.example.tsx', {
-    title: 'Custom label',
-    takeScreenshot: false,
-  })
+  .addExample('FileInput/story/CustomButtonLabel.example.tsx', 'Custom label')
   .addExample('FileInput/story/InitialValue.example.tsx', 'Initial value')
   .addExample(
     'FileInput/story/LimitedFileCount.example.tsx',