+++ packages/picasso/src/NumberInput/story/index.jsx
@@ -14,9 +14,6 @@ page
   .createChapter()
   .addExample('NumberInput/story/Default.example.tsx', 'Default')
   .addExample('NumberInput/story/Disabled.example.tsx', 'Disabled')
-  .addExample('NumberInput/story/Status.example.tsx', {
-    title: 'Status',
-    takeScreenshot: false,
-  })
+  .addExample('NumberInput/story/Status.example.tsx', 'Status')
   .addExample('NumberInput/story/WithIcon.example.tsx', 'With Icon')
   .addExample('NumberInput/story/Sizes.example.tsx', 'Sizes')