+++ packages/picasso/src/RatingStars/story/index.jsx
@@ -24,19 +24,16 @@ page
   .addExample('RatingStars/story/Default.example.tsx', {
     title: 'Default',
     description: 'Default behavior',
-    takeScreenshot: false,
   })
   .addExample('RatingStars/story/NonInteractive.example.tsx', {
     title: 'Non Interactive',
     description:
       'The rating can be used in non-interactive mode. This mode prohibits changing its value.',
-    takeScreenshot: false,
   })
   .addExample('RatingStars/story/Sizes.example.tsx', {
     title: 'Sizes example',
     description:
       'The rating component has two sizes - small (default) and large. It can be set via size parameter.',
-    takeScreenshot: false,
   })
   .addExample('RatingStars/story/CustomIcons.example.tsx', {
     title: 'Custom Icon',