+++ packages/picasso/CHANGELOG.md
@@ -1,5 +1,36 @@
 # Change Log
 
+## 26.3.2
+
+### Patch Changes
+
+- [#2970](https://github.com/toptal/picasso/pull/2970) [`ff4a70ef`](https://github.com/toptal/picasso/commit/ff4a70ef7b568a8a94131f30598577399c188cf3) Thanks [@iatanas0v](https://github.com/iatanas0v)! - ---
+
+  ### Autocomplete
+
+  - show options on every click
+
+## 26.3.1
+
+### Patch Changes
+
+- [#2959](https://github.com/toptal/picasso/pull/2959) [`984b826a`](https://github.com/toptal/picasso/commit/984b826af8f54835c6914586c338d60c64c741ff) Thanks [@OndrejTuma](https://github.com/OndrejTuma)! - ---
+
+  - bump `notistack` to `1.0.10`
+
+- Updated dependencies [[`984b826a`](https://github.com/toptal/picasso/commit/984b826af8f54835c6914586c338d60c64c741ff)]:
+  - @toptal/picasso-shared@9.0.1
+
+## 26.3.0
+
+### Minor Changes
+
+- [#2952](https://github.com/toptal/picasso/pull/2952) [`7185c71b`](https://github.com/toptal/picasso/commit/7185c71bfb7896d125c9cb2d795589f2ba8b8ba1) Thanks [@tiagoporto](https://github.com/tiagoporto)! - ---
+
+  ### Icon
+
+  - Add inbox icon
+
 ## 26.2.2
 
 ### Patch Changes